+++ app/src/main/java/com/techcourse/dao/template/UserDao.java
@@ -1,9 +1,5 @@
-package com.techcourse.dao;
+package com.techcourse.dao.template;
 
-import com.techcourse.dao.template.DeleteAllJdbcTemplate;
-import com.techcourse.dao.template.InsertJdbcTemplate;
-import com.techcourse.dao.template.SelectJdbcTemplate;
-import com.techcourse.dao.template.UpdateJdbcTemplate;
 import com.techcourse.domain.User;
 import java.util.List;
 import javax.sql.DataSource;
