+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,7 +1,7 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.SimpleUserDao;
-import com.techcourse.dao.SimpleUserHistoryDao;
+import com.techcourse.dao.simple.SimpleUserDao;
+import com.techcourse.dao.simple.SimpleUserHistoryDao;
 import com.techcourse.domain.User;
 import com.techcourse.domain.UserHistory;
