+++ app/src/test/java/com/techcourse/dao/SimpleUserDaoTest.java
@@ -3,6 +3,7 @@ package com.techcourse.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.techcourse.config.DataSourceConfig;
+import com.techcourse.dao.simple.SimpleUserDao;
 import com.techcourse.domain.User;
 import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
 import org.junit.jupiter.api.BeforeEach;
