+++ app/src/test/java/com/techcourse/service/MockUserHistoryDao.java
@@ -1,9 +1,8 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.dao.template.UserHistoryDao;
 import com.techcourse.domain.UserHistory;
 import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.core.JdbcTemplate;
 import javax.sql.DataSource;
 
 public class MockUserHistoryDao extends UserHistoryDao {
